backend/requirements.txt
# API and server
fastapi>=0.95.2
uvicorn>=0.22.0
python-dotenv>=1.0.0
sse-starlette>=1.6.5

# Database
supabase>=2.3.4
sqlalchemy>=2.0.29
# Lock dependency versions that work with Supabase
httpx>=0.24,<0.26
gotrue<2.9

# LLM and conversation flow
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
portkey-ai==1.10.2
langsmith==0.2.11
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-settings==2.7.1
langgraph>=0.0.39

# Utilities
requests>=2.28.2
